@@ -4,9 +4,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Rewrite;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Swashbuckle.AspNetCore.Swagger;
@@ -35,19 +35,32 @@ namespace api
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment  env, ILoggerFactory loggerFactory)
         {
-            app.UseMvc();
+            //app.UseMvc();
 
-            app.UseSwagger();
-            app.UseSwaggerUI(c =>
+            //app.UseSwagger();
+            //app.UseSwaggerUI(c =>
+            //{
+            //    c.SwaggerEndpoint("/swagger/v1/swagger.json", "Generate Random Data API V1");
+            //});
+
+            //var redirectRootToSwagger = new RewriteOptions()
+            //    .AddRedirect("^$", "swagger");
+            if (env.IsDevelopment())
             {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "Generate Random Data API V1");
-            });
+                app.UseDeveloperExceptionPage();
+            }
 
-            var redirectRootToSwagger = new RewriteOptions()
-                .AddRedirect("^$", "swagger");
-            app.UseRewriter(redirectRootToSwagger);
+            app.UseRouting();
+
+            app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
+            //app.UseRewriter(redirectRootToSwagger);
         }
     }
 }
